+++ R/pip_load_inventory.R
@@ -1,5 +1,12 @@
 #' Load inventory of welfare aggregate datasets
 #'
+#' @description `r lifecycle::badge("deprecated")`
+#'
+#' This function was deprecated because DLW data will loaded directly from a
+#' flat folder structure which allows bypassing the use of `datalibweb` in
+#' Stata, making the `pipdp` Stata package useless.From now on, use the `dlw`
+#' functions: `pip_load_dlw_inventory`, `pip_find_dlw`, and `pip_load_dlw`.
+#'
 #' @param inv_file character: file path to be loaded.
 #' @inheritParams pip_inventory
 #' @inheritParams pip_create_globals
@@ -12,11 +19,15 @@
 #' \dontrun{
 #' pip_load_inventory()
 #' }
-pip_load_inventory <- function(root_dir          = Sys.getenv("PIP_ROOT_DIR"),
-                               maindir           = pip_create_globals(root_dir)$PIP_DATA_DIR,
+pip_load_inventory <- function(root_dir = Sys.getenv("PIP_ROOT_DIR"),
+                               maindir  = pip_create_globals(root_dir)$PIP_DATA_DIR,
                                inv_file = fs::path(maindir,
                                                  "_inventory/inventory.fst")
                                ) {
+
+  lifecycle::deprecate_soft("0.1.13",
+                            "pip_load_inventory()",
+                            "pip_load_dlw_inventory()")
 
   #--------- Load Data ---------