+++ NEWS.md
@@ -1,3 +1,11 @@
+# pipload (development version)
+* read most recent dir in each object of list `create_dir()` inside `pip_create_globals()`
+
+* Create `pip_load_dlw_inventory()`, `pip_find_dlw()`, and `pip_load_dlw()`
+
+* add `lifecycle` badges
+
+
 # pipload 0.1.12
 * Allow "all" in `pip_load_all_aux()`
 
@@ -11,7 +19,7 @@
 # pipload 0.1.10
 * make sure `pipload` work in RS connect server. 
 
-* `pip_create_globals()` vignnete.
+* `pip_create_globals()` vignette.
 
 # pipload 0.1.9
 * add max years for countries and aggregates in `pip_create_globals()` this